isolating document scanner to only android
